requirements.txt
django
djangorestframework
django-cors-headers
django-extensions
psycopg2-binary
ippanel
pillow